SKALE-1947 refactoring
